refactor: 카카오 공유 스크립트 분리
