+++ src/App.tsx
@@ -9,6 +9,7 @@ import { analytics } from "@/analytics/analytics";
 import ConfirmAlert from "./components/alert/ConfirmAlert";
 import { Toaster } from "./components/ui/toaster";
 import PrayCardCreatePage from "./pages/PrayCardCreatePage";
+import KakaoInit from "./components/share/KakaoInit";
 
 const App = () => {
   useEffect(() => {
@@ -70,6 +71,7 @@ const App = () => {
       </div>
       <Toaster />
       <ConfirmAlert />
+      <KakaoInit />
     </div>
   );
 };
